+++ fuel-vm/src/interpreter/executors/instruction/tests/reserved_registers.rs
@@ -243,7 +243,6 @@ fn writes_to_ra(opcode: Opcode) -> bool {
         Opcode::CFS => false,
         Opcode::ECAL => true,
         Opcode::BSIZ => true,
-        Opcode::BLDC => false,
         Opcode::BLDD => false,
     }
 }
@@ -361,7 +360,6 @@ fn writes_to_rb(opcode: Opcode) -> bool {
         Opcode::CFS => false,
         Opcode::ECAL => true,
         Opcode::BSIZ => false,
-        Opcode::BLDC => false,
         Opcode::BLDD => false,
     }
 }